// uniquely/form-registration-pwa/src/type.ts
import type {StringifyableRecord} from '@alwatr/type';
import type {IconBoxContent} from '@alwatr/ui-kit/card/icon-box.js';
import type {TopAppBarContent} from '@alwatr/ui-kit/top-app-bar/top-app-bar.js';

export interface FormData extends StringifyableRecord {
  formId: string;
  name: string;
  phoneNumber: string;
}

export interface BoxType extends IconBoxContent {
  wide?: boolean;
  slot?: string;
}

export interface PageHomeContent extends StringifyableRecord {
  topAppBar: TopAppBarContent;
  boxList: Array<BoxType>;
}